fix: [Reservation/Event] 코드 리뷰 지적 사항 수정 (#49) (#216)
- totalSeats 타입 Int → Long 변환 (SeatDto, EventServiceClient, ReservationDto)
- SeatService.getSoldSeatIds() .toInt() 제거로 오버플로우 방지
- ReservationService.getSeatStatus() sold/hold 집합 중복 제거 후 가용 좌석 계산
- EventServiceClient Resilience4j 적용 (CircuitBreaker, Retry, TimeLimiter, FallbackFactory)
- ReservationController KDoc 오귀속 수정 (REQ-RSV-003 ↔ REQ-RSV-001)
- InternalSeatControllerTest totalSeats JSON 필드 검증 추가
- SeatServiceTest totalSeats 독립 값으로 분리 (sold 수량과 결합 제거)

+++ backend/event-service/src/main/kotlin/com/ticketqueue/event/dto/SeatDto.kt
@@ -50,7 +50,7 @@ class SeatDto {
     data class SoldSeatsResponse(
         val scheduleId: UUID,
         val soldSeatIds: List<UUID>,
-        val totalSeats: Int
+        val totalSeats: Long
     )
 
     /** 내부 API 응답 - Reservation Service가 좌석 선점 시 스냅샷 저장용으로 사용 */
