+++ backend/event-service/src/main/kotlin/com/ticketqueue/event/service/SeatService.kt
@@ -123,7 +123,7 @@ class SeatService(
 
         val soldSeatIds = seatRepository.findSoldSeatIdsByScheduleId(scheduleId)
         val totalSeats = seatRepository.countByEventScheduleId(scheduleId)
-        return SeatDto.SoldSeatsResponse(scheduleId = scheduleId, soldSeatIds = soldSeatIds, totalSeats = totalSeats.toInt())
+        return SeatDto.SoldSeatsResponse(scheduleId = scheduleId, soldSeatIds = soldSeatIds, totalSeats = totalSeats)
     }
 
     /**
